fix: add a lib.js to export functions without running main
